refactor: :recycle: split BurgerMenu constants to separate folders
