+++ components/Navigation/BurgerMenu/BurgerMenu.tsx
@@ -1,20 +1,12 @@
 import S from './BurgerMenu.styles';
+import routePages from '@/config/routePages';
+import capitalizeWord from '@/utils/capitalizeWord';
 
 const BurgerMenu = ({ isOpen }: { isOpen: boolean }) => {
-	const pages = [
-		{ name: 'home', path: '/' },
-		{ name: 'contact', path: '/contact' },
-		{ name: 'about', path: '/about' },
-		{ name: 'texts', path: '/texts' },
-	];
-
-	const capitalizeWord = (word: string) =>
-		word.charAt(0).toUpperCase() + word.slice(1);
-
 	return (
 		<div>
 			<S.UnorderedList isOpen={isOpen}>
-				{pages.map((page) => (
+				{routePages.map((page) => (
 					<S.ListItem key={page.name}>
 						<S.NextLink href={page.path}>{capitalizeWord(page.name)}</S.NextLink>
 					</S.ListItem>
