+++ config/routePages.ts
@@ -1,0 +1,8 @@
+const pages = [
+	{ name: 'home', path: '/' },
+	{ name: 'contact', path: '/contact' },
+	{ name: 'about', path: '/about' },
+	{ name: 'texts', path: '/texts' },
+];
+
+export default pages;
