+++ utils/capitalizeWord.ts
@@ -1,0 +1,4 @@
+const capitalizeWord = (word: string) =>
+	word.charAt(0).toUpperCase() + word.slice(1);
+
+export default capitalizeWord;
